@@ -58,19 +58,19 @@
     "release": "bumpp && pnpm run publish"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^2.27.3",
-    "@types/node": "^18.19.87",
+    "@antfu/eslint-config": "^4.15.0",
+    "@types/node": "^18.19.112",
     "@types/vscode": "1.95.0",
-    "@vscode-use/utils": "^0.1.56",
+    "@vscode-use/utils": "^0.1.59",
+    "@vscode/vsce": "^3.5.0",
     "bumpp": "^9.11.1",
-    "eslint": "^8.57.1",
+    "eslint": "^9.29.0",
     "esno": "^0.16.3",
     "find-up": "^7.0.0",
-    "lazy-js-utils": "^0.1.43",
+    "lazy-js-utils": "^0.1.45",
     "tsup": "^6.7.0",
     "typescript": "^5.8.3",
     "vitest": "^0.29.8",
-    "vsce": "^2.15.0",
     "yamljs": "^0.3.0"
   }
 }
